requirements.txt
coverage==7.6.1
iniconfig==2.0.0
model-bakery==1.19.5
pluggy==1.5.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.9.0
sentry-sdk==2.14.0
django-debug-toolbar==4.4.6
asgiref==3.8.1
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.0.3
